# api/open-api-spec.yml
openapi: 3.0.0
info:
  title: Awesome Privacy API
  description: API for accessing information on privacy-focused services.
  version: 1.0.0
servers:
  - url: https://api.awesome-privacy.xyz
paths:
...
  /search/{searchTerm}:
    get:
      summary: Returns an array of listings matching the given search term
      parameters:
        - in: path
          name: searchTerm
          required: true
          schema:
            type: string
          description: The term to search for
      responses:
        '200':
          description: A list of services matching the search term
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Service'
        '404':
          description: Search term not specified
...
components:
  schemas:
    ShortService:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        url:
          type: string
      required:
        - name
        - description
        - url
    Service:
      allOf:
        - $ref: '#/components/schemas/ShortService'
        - type: object
          properties:
            github:
              type: string
            icon:
              type: string
            followWith:
              type: string
            securityAudited:
              type: boolean
            openSource:
              type: boolean
            acceptsCrypto:
              type: boolean
            tosdrId:
              type: number